Add support for reading mroutes and mgroups from a configuration file.
This changeset adds support for joining multicast groups and setting
multicast routes, only IPv4 so far, from a config file.

   mgroup from IFNAME group MCGROUP
   mroute from IFNAME source ADDRESS group MCGROUP to IFNAME [IFNAME ...]

By default the file /etc/smcroute.conf is read, but it is also possible
to specify a different file using the -f FILE command line argument.

An example smcroute.conf is included, it is not part of the distribution,
but should perhaps be included in /usr/share/doc/smcroute for reference.

Further, a complete overhaul of the man page and the online -h help text
have been done. The AUTHORS file received a minor update and a TODO file
was also added.

Before first public release we should support signals (SIGHUP!), fix
parser to support IPv6, and support user supplied conf file.