Revert 8c7b4c0 due to unmotivated size increase.
Jean-Baptiste Maillet brought to my attention the unmotivated increase
in binary size and (comparatively) large external dependency introduced
in commit 8c7b4c0:

Event Loop
----------

$ size src/smcroute
   text	   data	    bss	    dec	    hex	filename
  30761	   1056	   2312	  34129	   8551	src/smcroute

$ size /usr/lib/libev.so.4.0.0
   text	   data	    bss	    dec	    hex	filename
  47696	    976	   2888	  51560	   c968	/usr/lib/libev.so.4.0.0

Even with libev built-in, the total numbers are:

$ size src/smcroute
   text	   data	    bss	    dec	    hex	filename
  78093	   1344	   5160	  84597	  14a75	src/smcroute

Significantly larger than without libev.

Without Event Loop
------------------

$ size src/smcroute
   text	   data	    bss	    dec	    hex	filename
  29605	   1028	   1896	  32529	   7f11	src/smcroute

The real motivation for 8c7b4c0 was to simplify the core of SMCRoute,
but the overhead in size cannot be motivated, hence this revert.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>