Fix hidden problem with dynamic interfaces, thanks to cppcheck.
The clever cppcheck tool alerted to this:

    [smcroute.c:178]: (style) Variable 'iface' is assigned a value that
    is never used.

Turns out there was unfinished code that was supposed to check if
inbound traffic for (*,G) handling was from an unknown interface.

Add check for unknown interface, possibly dynamically added after
smcroute was started.  We don't support such interfaces right now,
but it could easily be added.

For some people it could be useful to instead exit on this error,
and let a process supervisor restart smcroute so it reenumerates
and picks up that dynamic interface.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>