Last attempt to fix .travis.yml for Coverity Scan
I get the following message from Travis when trying to trigger a scan,
looks like a broken tarball?  My guess is that I've managed to set the
wrong security token so the tarball is scrambled.  This is the last
attempt to fix it by diffing with my other projects and verbatim paste
from the Coverity project page ...

    Coverity Scan configured to run on branch master
    Coverity Scan analysis authorized per quota.
    Downloading Coverity Scan Analysis Tool...
    2014-09-20 09:30:24 URL:https://scan.coverity.com/users/sign_in [8576] -> "/tmp/cov-analysis-Linux.tgz" [1]
    Extracting Coverity Scan Analysis Tool...
    /tmp/coverity-scan-analysis ~/build/troglobit/smcroute
    gzip: stdin: not in gzip format
    tar: Child returned status 1
    tar: Error is not recoverable: exiting now

  -- https://travis-ci.org/troglobit/smcroute/builds/35798465

If it still doesn't work, then I'm going to assume there's something
wrong in the interface between Travis and Coverity.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>