Make output of build id optional
With free software, anyone can inspect the source code for malicious flaws. But
the software is often provided as binary packages to its users. The idea of
“deterministic” or “reproducible” builds is to empower anyone to verify that no
flaws have been introduced during the build process by reproducing
byte-for-byte identical binary packages from a given source.
See also: https://wiki.debian.org/ReproducibleBuilds

Prior to this change a timestamp generated at build time is included in the
output of 'smcroute -v'. As a result the binary is unreproducible.

With this change, users that still want 'smcroute -v' to include a build id can
run './configure --enable-build-id' to get the build id back (or even specify a
custom build id).