Avoid storing generated files in GIT
This patch removes generated files from the GNU Configure & Build system,
autoconf and automake.  Hence no more configure script readily available.

A developer needs to have autoconf, at least v2.61, and automake, at least
v1.11, to call on the autogen.sh script to generate the necessary files.

When a release is made autoconf et al generates a distribution archive
with all the required files, configure, Makefile.in, etc.

This patch affects nothing in the release process or the files an end user
needs.  It only reduces the overhead of files stored in the VCS.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>