Fix issue #14: Add support for executing a script on reload/install
This patch adds support for calling an external script at startup and
reload (SIGHUP) as well as when (dynamic) source-less multicast routes
are installed.

    ./smcroute -s /path/to/script

The script is called when SMCRoute has started up, or has received
SIGHUP and just reloaded the configuration file, and when a new
source-less rule have been installed.

The script is called without command line arguments but with up to
two environement variables set:

- source: The source IP address of the multicast group sender
- group:  The IP multicast group address

The environment variables are only set when a "source-less" route is
installed in the kernel.

To protect the communication between SMCRoute and the kernel all files
and sockets are opened with CLOEXEC set.  So when the script runs it
does not have access to any file descriptors or sockets from SMCRoute.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>