Fix issue #14: Add support for executing a script on reload/install
This patch adds support for calling an external script 'on reload'
and/or 'on install' of (dynamic) source-less multicast routes.

The syntax for `/etc/smcroute.conf` looks as follows:

    exec /path/to/script [optional params] [on <reload | install>]

The 'on reload' condition is executed when SMCRoute has started up and
installed all static rules.  The 'on install' condition is for when a
new source-less rule is triggered.

The command can be given in many different forms, even on multiple
lines.  If the same line is repeated only the last line, for a given
condition, is used.  It is possible to use the same script both when
SMCRoute signals that all static routes have been installed and when
SMCRoute signals a single source-less rule have been triggered.  In
the latter case SMCRoute sets two environment variables ("source"
and "group"), which the script can use to distinguish between the
two use-cases.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>