Replace implicit exit(255) in smclog() on LOG_ERR with explicit exit(255)
The `smclog()` code has always been a bit of a hairy mess.  This is the
first commit in a series attempting to clean up and simplify it.

This patch replaces the implicit `exit(255)`, when `smclog()` is called
with severity `LOG_ERR`, with actual `exit(255)` in each place that
calls `smclog()` with `LOG_ERR`.  It is slightly more code, but a lot
clearer to the casual reader.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>