Simplify daemonizing and cleanup code (SEMANTIC CHANGE)
Previously it was possible to start the smcroute daemon *and* send
client commands to it, all in one go.  This caused overly complex
code and a local re-implementation of the standard `daemon()` API.

This patch simplifies this code path, cleans up and significantly
improves the readability of the `main()` function by sacrificing the
above mentioned use-case.

Hence, after this patch it is no longer to call SMCRoute like this:

    smcroute -d -a eth0 192.168.1.1 225.1.2.3 virbr0

The man page has been updated accordingly.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>