Remove bogus interface filtering for INADDR_ANY
Align the IPv4 VIF setup with what is already done for IPv6 VIFs.
Static multicast routing do not need an IP address on any of the
inbound or outbound interfaces.  All that is needed is to filter
out `IFF_LOOPBACK` and those interfaces without `IFF_MULTICAST`

A later fix will address the fact that SMCRoute currently use
getifaddrs(), which does not return DHCP/PPPoE interfaces that
haven't yet received a lease.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>