Fix #32: Add ttl-threshold for outbound interfaces
This patch adds support for configuring administrative TTL scoping of
multicast.  I.e., the TTL of a multicast stream must exceed a given
threshold to be allowed to be routed.  Useful for filtering out any
local-only multicast from being forwarded on specific interfaces.

Syntax:

    phyint IFNAME ttl-threshold <1-255>

The default threshold for all enabled VIFs/MIFs is 1.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>