Acknowledge client SIGHUP by touching the pidfile
This patch gives SMCRoute the ability to acknowledge a received SIGHUP
by touching its pidfile, /run/smcroute.pid.  A handy feature for both
regular users and process supervision systems like Finit when peforming
a reconfiguration of a system.  Simply comparing the mtime of the config
file and the pid file will determine if a change "took".

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>