Check user:group exists already in arg parser
When the user requests to drop privs, check the given user:group before
we try to start the server.  This way we only need to deal with uid/gid
in the rest of the code *and* we can let the pidfile() function chown()
the PID file to the correct user.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>