Support for IPv4 (*,G/LEN) routes
This patch adds initial support for IPv4 (*,G/LEN) routes as proposed in
GitHub issue #50.  The recommended way to use this is in smcroute.conf:

    mroute from eth0 group 225.0.0.0/24 to eth1 eth2

This will match all inbound multicast in the 225.0.0.0 - 225.255.255.255
range, from eth0, and add SSM kernel routes to eth1 and eth2.

The same can be done using the client interface:

    smcroute -a eth0 0.0.0.0 225.0.0.0/24 eth1 eth2

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>