Support unlimited number of network interfaces
Remove the limit on the number of system network interfaces (max. 40) that
may be handled by smcroute. Now the number of network interfaces is only
limited by the available RAM.

Note that this does not remove limit on the number of network interfaces
actually used for multicast routing as that is a kernel limit (typically
32 VIFs max.)

Signed-off-by: Martin Buck <mb-tmp-tvguho.pbz@gromit.dyndns.org>