Always add MFC entry in response to kernel cache miss
If there is no suitable dynamic MC route defined for a cache miss reported
by the kernel, add an MFC entry with an empty list of egress interfaces
instead of doing nothing at all.  If we wouldn't do anything, the kernel
would keep the provisional (unresolved) MFC entry it created automatically
when receiving the packet that caused the cache miss for 10s.  However,
there is a kernel limit on at most 10 such unresolved MFC entries at a time,
so receiving more than 10 MC streams without matching routes might cause
additional MC streams with suitable dynamic routes defined to be ignored.
This is because the kernel will not report a cache miss for them if it is
unable to create a new unresolved MFC entry.  By always creating (resolved)
MFC entries, we keep the number of unresolved entries low, so the limit is
not reached and the kernel may always report cache misses for new traffic.

Note that this should be used in combination with periodic cache expiration
(option -c) to avoid acccumulating more and more MFC entries which never
get deleted again.

Signed-off-by: Martin Buck <mb-tmp-tvguho.pbz@gromit.dyndns.org>