Refactor, introduce mini-event loop with callback per socket
This patch changes the create_socket() API, used in all subsystems, to
take an optional callback.  The socket module can then register each
socket and callback in a list, which it then can traverse for polling.

The resulting code has a greater degree of separation and delegates all
functionality to each submodule, instead of being spread out.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>