Check for clock_gettime(), timer_create(), et al in -lrt + -lpthread
When building SMCRoute statically some C libraries need -lpthread, due
to how they've implemented timer_create().  Most of these APIs are also
in -lrt, so that's a potentialk dependency for all of them.

The check is implemented so that smcroutectl is *not* linked with them,
since it does not have such dependencies.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>