Add (S,G) to internal linked list *after* successful write to kernel
When the kernel receives multicast from an unknown source we may add a
multicast route if a matching (*,G) rule exists.  For bookkeeping we
maintain a list, which can be viewed with `smcroutectl show`.  This
patch fixes a small ordering bug which always added to the list, even
if setting the route in the kernel failed.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>