smcroute: Simple compat wrapper script
In the Debian package there is a /etc/smcroute/startup.sh script which
some users may still be relying on.  Instead of attempting to convert
that script we provide this simplistic wrapper.

The idea is that the init script starts smcroute, the daemon, and then
calls startup.sh to add routes and join groups.  Should be good enough.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>