Make sure to register alias interfaces (baseif:num) as baseif
Linux alias interfaces are actually implemented on top of the
base interface.  They share the same MAC address and all the
interface counters are on the base interface.  So when we get
an alias interface from getifaddrs() we should register it as
the base interface so we don't end up with two VIFs fro the
same physical interface.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>