Fix #106: Patch timer nsec handling
When using timers, a small window exists where they can end up in a non
running state.  This is caused by how the nsec value in struct timespec
is handled when restarting the timers.

When a timer is less than a second from expiring it is not restarted
correctly, resulting in timers that never expire.

Signed-off-by: Jonas Holmberg <jonas.holmberg@westermo.se>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>