test: initial unit test framework, requires docker
This patch adds the initial unit test framework for SMCRoute.  It is
built around Docker and CAP_NET_ADMIN to allow for creating network
interfaces and set up routing as a non-root user.

To run the test(s) outside of GitHub Actions:

    ~$ docker pull ghcr.io/troglobit/misc:latest
    ~$ docker run --cap-add=NET_ADMIN -v`pwd`:`pwd` -w `pwd` -it ghcr.io/troglobit/misc

The misc:latest container is currently based on Alpine 3.13 and comes
with iproute2, tcpdump, and some other useful things already in place.
This is also what the GitHub Actions use to build SMCRoute.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>