test: use /tmp for temporary files
This changes all tests, except isolated.sh, which needs more extensive
changes, to use temporary files of the for "/tmp/$NM/conf".  This is
later used by the teardown method for cleaning up.

Main reason for change is that when running `unsare -mrun` as root we
no longer have write capability to the checked out repository, so the
tests always fail when running in a docker container.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>