.github: build: use sub-sub-directory for build and test
We'd like to run make distcheck, but that won't work (see below), so
we emulate it by testing `make dist`, unpacking one of the archives,
and testing build in a sub-sub-directory.

Also, allow for build directory to be writable by nobody, since our
`make check` runs in an `unshare -mrun` that drops root privs.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>