Refactor, generalize struct mroute4 with inet_addr_t source + group
This is a massive refactor, thankfully the unit tests saved me early on
from a few mistakes.  Nevertheless, there may very well be regressions
introduced in this commit!

Overall though, most of the code changes turned out for the better and
actually improved readability.  A few more iterations and we may soon be
able to collapse a few ipv4/ipv6 blocks into a single block :)

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>