.github: coverity: seed caching with coverity .md5
We like to always run the latest coverity-scan tool, but we don't like
to download the tarball every time.  Coverity offer a .tar.gz.md5 file
that can be used to check if the latest download.

This patch adds a step to the job that fetches that .md5 file and also
uses the contents as the cache + restore key.  With fallback to the old
key we've used prior.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>