Refactor, relocate inet.[ch] content to mrdisc.[ch] to free prefix
We want to use the inet_ prefix for our housekeeping IPv4/IPv6 wrapper
functions.  So this patch relocates the contents of inet.c to mrdisc.c,
which was only ever the only user of this functionality.  Any future
support for IPv6 will have to do the next level of refactoring.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>