kern:group_req() - fix struct ip_mreq fallback handling
This code is the fallback in case of missing RFC 3678 support.  It was
previously refactored but never tested.  It had no way of ever working
since it wasn't complete, at all.  Now it at least treats S,G joins as
*,G join for both IPv4 and IPv6.

This should even work on *BSD systems, although that's not been tested.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>