Minor correction, VIF is an *unsigned* type, at most 16 bits
The current kernel limitation is MAX 32 simulataneous VIFs (and MIFs),
in some parts of the kernel code it's 8 bits and in newer parts 16 bits
but alsways unsigned.

For the few that have patched the kernel MAXVIFS define to something
greater than 127 this bug fix may be relevant.  To all others it's just
a NOP.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>