mroute.c: add upcall header validation to IPv4 and IPv4 NOCACHE
Add basic validation of igmpmsg (IPv4) and mrtctl (IPv6) to prevent
bogus messages from being dropped and logged.  In particular this
applies to IPv6 since the socket gets all application level MLD
messages.  We are only interested in the upcalls for (*,G) routing.

This patch also renames the local variables used for kernel upcalls to
follow the convention used in other structs.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>