Drop configure support for --disable-config and --disable-client
The overhead for supporting either of them is apx. 4 kiB in the
resulting stripped binary smcrouted.  I have decided it's not worth
maintaining the complexity in the code to meet these two external
requirements.

The requirements initially came from users of the old SMCRoute that
wanted to keep their resulting firmware images small, and also didn't
need the .conf support.

As things stand, I have to weight external requirements from users
against the burden of maintaining complexity.  Since nobody is paying
for this, I'm ripping it out so I can share code between the IPC and
.conf parsers instead, and possibly save a few bytes that way.

There is a lot of of angst associated with removing features.  I expect
a backlash from this and suspect a few users will be very angry.  Then
again, if you want to drive an open source project in certain direction,
pick up a shovel and help out.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>