Initial support for more graceful SIGHUP/reload
This extensive patch adds support for updating existing multicast routes
without disturbing existing flows.  Not only routes, but also groups and
phyint to vif mappings are handled in the same manner.

I.e., when SIGHUP or restart command is received, instead of closing the
multicast routing sockets (and all the multicast group sockets) and thus
dropping existing flows, mroutes, mgroups, and phyints are marked with a
'unused' flag, the .conf file is re-read and all entries that remain the
same, or updated (e.g., more/less outbound interfaces) have the 'unused'
flag cleared, and are possibly updated.  Remaining entries are deleted.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>