Fix #103: add support for including other .conf files
This patch adds support to the .conf parser for including any pattern of
file that glob() can understand.  For insstance:

    include foo.conf
    include /etc/smcroute.d/*.conf

Thanks to the new diff engine in smcrouted, it's now possible to drop a
route in /etc/smcroute.d/foo.conf send SIGHUP and have it installed in
a flash.  Removing the route is equally simple, remove the file and do
another SIGHUP.  No other established flows should be affected.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>