test: new test, poison.sh verifies stop-filter routes
This test verifies that stop-filter, or "poison pill", routes work as
intended.  I.e., unknown inbound multicast is blocked and only such
that is actually wanted is properly routed.

Example:

    mroute from eth0 group 225.1.2.3 to eth2
    mroute from eth1 group 225.1.2.3 to eth2

These two multicast routes are dynamically installed in the kernel MFC
when any source originating from eth0 or eth1 with a matching group is
interecpted by the kernel and smcrouted is notified (NOCACHE msg).

If multicast to the same multicast group comes in on eth2, smcrouted
adds a stop-filter route, i.e., a route with no outbound interfaces, to
prevent further NOCACHE upcall messages from the kernel.  This does not
affect any flow in the intended direction, esatablished before or after
the stop-filter is created.

However, should source `S` from eth0 suddenly appear on eth2, this would
be considered a WRONGVIF event, which is not handled by smcrouted.  This
typically occurs when a layer-3 topolgy change is taking place.  For the
moment, users are recommended to either try pimd/pim6sd, or add a switch
with multicast snooping between eth2 and the rest of the network.  When
a snooping switch is active it only forwards multicast to eth2 when it
has sent a join (mgroup) for a given (source and) multicast group.

Issue #143

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>