test: increase timeout for checking (*,G) cache flushing
Turns out increasing the tshark wait delay caused expire.sh to start
failing.  This was caused by overly optimistic timeouts in the test,
which this patch adjusts.

Also, schedule the test to start first since it's taking a bit longer
now to complete.  With `make -j9 check` this means the total test time
should not be that bad.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>