test: new test, basic multicast-over-GRE-tunnel
The kicker is to have the reverse path set up properly.  I.e., the
routes from each router back to the source of the multicast.  If
that's not set up, the kernel on R2 will not generate a NOCACHE
message to SMCRoute so it installs a matching (S,G) route.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>