When removing a (*,G) route, check for overlap w/ known (S,G) route
All (*,G) routes, incluing (*,G/LEN) and (S/LEN,G[/LEN]), may spawn
kernel (S,G) routes on inbound traffic.  It may also overlap with
existing configured (S,G) routes, so when deleting a (*,G) route, or
even just modify its outbound interfaces, we need to ensure that we
transition back to a state that is the logical OR with the configured
(S,G) route.

This patch adds a new function, do_mroute_del_ssm_check() that helps
with coalescing and updating the resulting kernel route.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>