Refactor, use TAILQ instead of LIST, new entries added last
This refactor is purely to improve the user experience; a new entry is
expected to be added last in a list.  So instead of sorting on output,
we switch to a TAILQ list.

Note, this required another sub-refactor in mroute_del_route(), which
unfortunately is a bit too involved.  I've tested the intricte parts
manually and run the automated tests with no obvious regression.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>