Massive refactor of mroute.c
This is basically a rewrite of the multicast route handling engine.  The
amount of internal lists to keep track of all routes has been reduced
from three to two, where one of them is the kern_list, which basically
just tracks what's been set in the kernel.  Similar to how mcgroup.c
works already.

The kern_list consists of two types of entries; stop-routes and regular
routes.  The latter are managed by two helper functions mfc_install(),
and mfc_uninstall().  They ensure that routes in the conf_list, which
may overlap, are coalesced before being sent to the kernel MFC.

This extra abstraction layer simplifies the rest of the code a great
deal, allowing for the removal of a lot of old fragile pieces as well
as scary code duplication.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>