Adjust support for adding *,G/LEN routes at runtime
This is tricky for, in particular, picking up existing inbound traffic
that currently is blocked.  We need to iterate over all installed MFC
entries that match the new *,G/LEN and update each.

Similar on removal, only in reverse, and still keeping OIFs for any of
the previously set routes that map to the same MFC entries.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>