If interface has changed ifindex on reload, delete any old VIF
On reload, if an interface with a known name has changed ifindex, it has
been removed and re-added with the same name.  We need to remove any VIF
that exist.  The kernel has probably already done this cleanup, so it's
mostly for our bookkeeping so we don't continue using stale VIFs.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>