Prune removed interfaces from existing routes
When an interface is either removed (phytint disable) or lost at runtime
we need to update all routes accordingly.  The kernel does not clean up
anything for us for free.

For routes with a lost inbound, we must drop the entire route, and for
routes with lost outbound we only prune the outbound.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>