Make sure add VIF functions can find "unsued" interfaces
On reload, and -N is given om the cmdline, we need to mark interfaces as
used when parsing `phyint foo enable` directives.  These functions must
skip the check for "unused" to be able to reset that flag.

On reload, and -N is *not* given, we mark the interfaces as unused as
soon as we find them in the list from getifaddrs().

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>