Prune VIFs from routes also when removing a VIF fails
The kernel does some housekeeping when an interface is deleted, so when
we show up to delete a VIF it may already be removed.  In such cases we
still want to clean up around our own entrance, so save return code and
do the cleanup before returning from mrouteY_del_Xif().

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>