Fix #167: cannot remove routes with `smcroutectl del`
This patch fixes the problem when removing a route that was previously
added using the command line tool smcroutectl.  Simply adding a route
and then deleting it did not work.  The test mrcache.sh reproduces the
issue and now ensures no future regressions.

Removing routes by deleting a line from the .conf file is not affected.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>