Avoid trying to delete inactive VIFs
When probing interfaces at startup, there's a check for IFF_MULTICAST,
if this flag is not set we try to delete its corresponding VIF/MIF.
This is for hanlding .conf reload scenarios where an interface has had
its MULTICAST flag dropped.

However, when starting up on Linux systems, the loopback interface has
no MULTICAST flag set.  This leads to the following bogus warning:

   Failed deleting VIF for iface lo: Resource temporarily unavailable

This patch makes sure to check if we have a registered kernel VIF/MIF
for an interface before attempting to delete it.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>