Fix #175: mrdisc_deregister() fails for never started vif
When first starting up SMCRoute with multiple interfaces, where at least
one phyint does *not* have mrdisc enabled, smcrouted fails with "parse
error" because mrdisc_deregister() fails to deregister a never before
registered mridsc interface.

This patch drops the bogus error, and adds a TODO entry for refactoring
the mrdisc support -- not just for this problem, but preparing for IPv6
and multi-platform support.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>