Drop repology status
This may have been a good idea, don't know what started it, but it does
not reflect well on the project itself.  It shows more red than green so
maybe the ones in the green feel good, but we as the upstream project do
not feel good and have no reason to promote other projects.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>